+++ packages/apps/esm-login-app/src/index.ts
@@ -1,7 +1,7 @@
 import { defineConfigSchema, getSyncLifecycle } from '@openmrs/esm-framework';
 import { configSchema } from './config-schema';
 import rootComponent from './root.component';
-import locationPickerViewComponent from './location-picker-view/location-picker-view.component';
+import locationPickerComponent from './location-picker/location-picker.component';
 import changeLocationLinkComponent from './change-location-link/change-location-link.extension';
 import logoutButtonComponent from './logout/logout.extension';
 
@@ -19,6 +19,6 @@ export function startupApp() {
 }
 
 export const root = getSyncLifecycle(rootComponent, options);
-export const locationPicker = getSyncLifecycle(locationPickerViewComponent, options);
+export const locationPicker = getSyncLifecycle(locationPickerComponent, options);
 export const logoutButton = getSyncLifecycle(logoutButtonComponent, options);
 export const changeLocationLink = getSyncLifecycle(changeLocationLinkComponent, options);
