+++ packages/apps/esm-login-app/src/location-picker/location-picker.resource.ts
