+++ packages/apps/esm-login-app/src/root.component.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Login from './login/login.component';
 import RedirectLogout from './redirect-logout/redirect-logout.component';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import LocationPickerView from './location-picker-view/location-picker-view.component';
+import LocationPicker from './location-picker/location-picker.component';
 
 export interface RootProps {}
 
@@ -12,7 +12,7 @@ const Root: React.FC<RootProps> = () => {
       <Routes>
         <Route path="login" element={<Login />} />
         <Route path="login/confirm" element={<Login />} />
-        <Route path="login/location" element={<LocationPickerView />} />
+        <Route path="login/location" element={<LocationPicker />} />
         <Route path="logout" element={<RedirectLogout />} />
       </Routes>
     </BrowserRouter>
